Update merge usn binary package logic
Currently, the binary packages we get from usns can have
a source package in them. This package is identified by
the name source in the binary package dict we operate
when merging all of the binary packages of the USNs.
When we are applying fixes in the system we check
if a affected package is in that merged binary package dict.
This will fail to find source packages, because the
affected packages will have the real name of the package
while the binary packages dict will have that package
named as source. We are now fixing this issue