tests: always build packages on a LXD container
When running integration tests with build_pr=1, we build the package
to be installed in an instance that will be used as a base for all
of the test scenarios. This package is build in a instance with the same
type as the one that will actually run the test. That running a vm test,
we will build the package in a vm. This is unnecessary, since we can
build the package in a LXD container, which is faster to create.
Because of that, all package builds will happen using a LXD container.